Multiplier is a decimal...

+++ app/models/transaction_log_work_stats.rb
@@ -12,7 +12,7 @@ class TransactionLogWorkStats
     @worker_id = nil
     @multiplier = 1
     @validated = false
-    
+
     TransactionLogEntry.where(product_id: product_id).
                         where('created_at <= ?', end_at).order(:created_at).each do |entry|
       if entry.work_id == work_id
@@ -27,12 +27,12 @@ class TransactionLogWorkStats
           @validated = true
 
         when 'multiplied'
-          @multiplier = entry.value.to_i
+          @multiplier = entry.value.to_d
         end
       end
     end
   end
-  
+
   def validated?
     @validated
   end
